Events ändern

+++ HA1_Maximilian_Bosse_EventVw_EJB/src/main/java/de/eventverwaltung/event/dao/GenericDAO.java
@@ -45,7 +45,7 @@ public abstract class GenericDAO<T> {
 		}
 	}
 	
-	public T find(long entityId) {
+	public T find(int entityId) {
 		return em.find(entityClass, entityId);
 	}
